Speed bus works, channels are now zero-based everywhere (0-511, not 1-512) this might change in the future so that the user sees channels as 1-512 but qlc handles them always as 0-511.

+++ main/src/device.h
@@ -40,7 +40,7 @@ class Device : public QObject
   virtual void setName(QString);
   QString name() const { return m_name; }
 
-  int address() const { return m_address; }
+  unsigned short address() const { return m_address; }
 
   /* Device Class */
   DeviceClass* deviceClass() { return m_deviceClass; }
